+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - #193 - Checking and add uniqueness_restrictions in the database
 - #198 - Add sorting of payments by ID
 - #175 - Updated security
+- #204 - Add a Central Bank client
 
 ## [1.4.0] - 2.04.2024
 - #167 - Testing the conversion of Product review to productDTO rating